+++ src/lib/types/proposal.ts
@@ -10,6 +10,7 @@ export enum ProposalStatus {
   REJECTED = "Rejected",
   FAILED = "Failed",
   DEPOSIT_FAILED = "Inactive",
+  CANCELLED = "Cancelled",
 }
 export const zProposalStatus = z.nativeEnum(ProposalStatus);
 
@@ -69,6 +70,24 @@ export interface Proposal {
   isExpedited: boolean;
 }
 
+export interface ProposalParams {
+  minDeposit: Coin[];
+  minInitialDepositRatio: number;
+  maxDepositPeriod: string;
+  votingPeriod: string;
+  vetoThreshold: number;
+  quorum: number;
+  threshold: number;
+  // expedited
+  expeditedVotingPeriod?: string;
+  expeditedThreshold?: number;
+  expeditedMinDeposit?: Coin[];
+  expeditedQuorum?: string; // only in sei
+  // emergency - only in initia
+  emergencyMinDeposit?: Coin[];
+  emergencyTallyInterval?: string;
+}
+
 export interface ProposalDeposit {
   amount: Coin[];
   depositor: BechAddr;
@@ -89,6 +108,14 @@ export interface ProposalData extends Proposal {
   totalDeposit: Nullable<Coin[]>;
   version: string;
   votingTime: Nullable<Date>;
+}
+
+export interface ProposalVotesInfo {
+  yes: Big;
+  abstain: Big;
+  no: Big;
+  noWithVeto: Big;
+  totalVotingPower: Big;
 }
 
 export interface ProposalVote {